+++ syncforge/src/commonMain/kotlin/dev/syncforge/conflict/ConflictStrategies.kt
@@ -1,6 +1,5 @@
 package dev.syncforge.conflict
 
-import dev.syncforge.api.ExperimentalSyncForgeApi
 import dev.syncforge.entity.SyncedEntity
 
 object ConflictStrategies {
@@ -16,16 +15,4 @@ object ConflictStrategies {
     inline fun <reified T : SyncedEntity> merge(
         noinline block: MergeScope<T>.(local: T, remote: T) -> T,
     ): ConflictStrategy = MergeStrategy(block)
-
-    /** Adapts a legacy [ConflictResolver] to [ConflictStrategy]. */
-    @Deprecated(
-        message = "Migrate resolver logic to ConflictStrategies or a custom ConflictStrategy. Will be removed in 1.0.",
-        replaceWith = ReplaceWith(
-            expression = "ConflictStrategies.lastWriteWins()",
-            imports = ["dev.syncforge.conflict.ConflictStrategies"],
-        ),
-    )
-    @ExperimentalSyncForgeApi
-    @Suppress("DEPRECATION")
-    fun fromResolver(resolver: ConflictResolver): ConflictStrategy = LegacyResolverStrategy(resolver)
 }